Fix imports where possible
This commit fixes broken imports for classes that were shaken up by being
brought over from an entirely different project. There are still missing
imports, but those will be resolved later. This is because most of them are
from libraries that may or may not get to be used as a dependency for the
project.