Move original source files to /src-legacy/
I decided to move these files to this new folder since they are so old, and
originally were not designed to be their own API (as they were part of a much
bigger game engine project). This was making it confusing and difficult to get
the code to a state where it would even compile. I figured that for an API I
was already planning to completely overhaul, it was not worth the effort. As
a result, the plan is to now simply rewrite the code from scratch based on the
legacy files. Once a legacy file is no longer needed for reference, it will be
deleted and removed from the repository.