Create InputSystem interface
Unlike the original version of this API, the central input system will not be
based on a static class. The plan is to have an implementation of the
InputSystem interface which can be instantiated instead. This allows for
other implements or extension of the standard InputSystem (which will be coded
later on) to be created.