Create device classes
These classes are used to represent and handle any and all input devices. I
have decided to use this terminology this time around, so as to avoid the
the confusing terminology of "controller". In the original API, all input
devices were considered controllers, even the Keyboard & Mouse. This led to
some awkward implementations were methods such as getLeftAnalogXAxis() would
be implemented, despite the device not even having such a peripheral.