Create Action.java
The intent of this class is to allow for input listeners to listen for action
events, rather than button events. Buttons are bound to actions, meaning that
ten controllers can have a specific button bound to an action, and the only
thing that a listener has to do is check for the action. This is useful for
both implementing multiple types of controllers as well as configurable
bindings.