Create ActionProfile.java
This is a new version of ControllerActionProfile from Ardenus Input v1. At the
moment, this newer version only has cleaner code and internal optimizations. It
is likely that more about this class will change in the future. But for now,
the purpose of an action profile is to allow multiple controllers share action
bindings. When no longer needed, they can simply be detached from the profile.
Action profiles also account for different types of controllers. For example,
if an action has a XBOX and PS4 button bound to it, then it will only bind
those buttons to their respective controllers when they are attached.