Create ActionBinds interface for ActionProfile class
I decided to move the methods from the ActionProfile class into an interface
which it will now implement instead. The Controller class will implement this
interface in the future as it will have its own internal "root" ActionProfile
for action binding. This is due to a change in design where Controllers will
attach ActionProfiles, rather than ActionProfiles attaching Controllers. This
is to fix the design flaw where ActionProfiles would "attach" Controllers by
simply copying over action bindings and unbindings. This was due to cause a
state management nightmare in the future, especially with controllers being
able to assign their own unique bindings without a profile.