Create Button.java and UIButton.java
I meant to add these earlier but forgot to commit them. So, this commit
fixes some broken references throughout the repository. There will probably be
a few more commits like this in the near future. The Button class acts as a
base class for other buttons in this UI. It doesn't contain any code at the
moment, but it may in the future. Currently, it's mainly just for OOP.
The UIButton class is acts as a way for buttons to exist in a user interface.
Code to interact with UIButtons will be added later.