Add update() method to InputSystem
Most if not all forms of input will require continous polling for user input
data, so this was added to the base interface for input systems. There are two
ways to call update(). One way is to pass a delta which represents how long in
milliseconds since the last update was called. Another way is to pass -1L for
the delta. If this is done, the input system implementation may either throw an
error or automatically determine how much time has passed since the last update
call. The standard input system implementation does the latter.