Add ButtonPresent.java
This annotation is to be used by future controller classes, which will annotate
fields of the type ControllerButton with this annotation. This annotation shall
indicate that a ControllerButton is present on a controller. An example of this
would be the A, B, X, and Y buttons on an XBOX controller.