Add ControllerButton.java
This is the base class for any buttons present on a controller. This slipped by
me when getting this repo's commits up to date. Had I saw this earlier, I would
have added both ButtonPresent and ControllerButton in a single commit. However,
it's almost 1:30AM and I am too tired to learn how amending commits works.