Add UIButtonSet.java
This will be implemented later. Its intended purpose is to act as a way to
group together UIButtons in a set. The basic concept is that UIButtonSets can
be enable/disable a group of buttons all at once. The goal is to make it less
tedious to implement user interfaces with more complicated inputs.