Removed GLFW based button and controller classes
I have removed these as implementing controllers this way (having them extend
off of a certain "provider" sort of class) undoes the work the API intends to
achieve. The goal of the API is to allow both entirely new controllers or
implementations for existing controllers to be usable. By having an XBOX
controller be considered a GLFW based controller, that makes it impossible
for the same type of controller to be implemented through another system (such
as XInput, for example).